fix: car name LIst 로 변경 및 예외처리 추가, 기존 코드 backing properties 사용하도록 수정
